Some new tech and a compile fix >Updated Manifest >Added a damage resistance aura >Added a periodic heal aura, for onDamageTaken() >Added a smart mana aura >Added onManaRestored(), onManaSpend(), and onPercentageMana() events to auras >Added mana, health, and special regen modifiers >Nerfed Bulwark, reduced rejuvenate >Added a statuseffect "last stand"

+++ TerraeRasa/TerraeRasa/src/entities/EntityLiving.java
@@ -242,7 +242,7 @@ public class EntityLiving extends Entity
 	 * @param showWorldText whether or not to display world text
 	 * @return true if the entity was healed, otherwise false
 	 */
-	public boolean heal(World world, int h, boolean showWorldText)
+	public boolean heal(World world, double h, boolean showWorldText)
 	{
 		if(health < maxHealth)
 		{
